(extension) fix prod build: keep standalone manifest JSON out of extension bundle
The extension build's crxjs popup pass cannot resolve the dango-manifests
package-subpath JSON that the standalone handler imported (the background pass
and the standalone build both can), which failed the prod build. Move the
bundled-spec resolution into standaloneManifestSpecs and load it via a
VITE_STANDALONE-gated dynamic import so the extension build dead-code-eliminates
it, while the standalone build still bundles it.
